refactor: extract PositionRowItem from TableContent
Split the ~400-line position row rendering logic from TableContent
into a dedicated PositionRowItem component to improve readability.
